Refactor user replace and delete to a job
